+++ app/models/utility/server.rb
@@ -2,6 +2,6 @@
 
 module Utility
   class Server
-    VERSION = "0.54"
+    VERSION = "0.55"
   end
 end